microservices/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

google-cloud-pubsub==2.18.1
google-auth==2.27.0

psycopg2-binary==2.9.9

pydantic==2.6.1
python-dotenv==1.0.1
